Implementar image upload

+++ Backend/src/main/java/com/oceanwinds/product/controller/ProductController.java
@@ -3,60 +3,66 @@ package com.oceanwinds.product.controller;
 import Global.dto.MessageDto;
 import Global.exceptions.AttributeException;
 import Global.util.PaginatedResponse;
-import com.oceanwinds.location.entity.Location;
 import com.oceanwinds.product.entity.Product;
 import com.oceanwinds.product.service.ProductService;
 import com.oceanwinds.product.entity.dto.ProductDto;
 import com.oceanwinds.feature.repository.FeatureRepository;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+
+
 
 @RestController
 @RequestMapping("/api/")
 @CrossOrigin(origins = "*")
 public class ProductController {
 
-    @Autowired
-    private ProductService productService;
+    private final ProductService productService;
+
+
+    private final FeatureRepository featureRepository;
 
     @Autowired
-    private FeatureRepository featureRepository;
+    public ProductController(ProductService productService, FeatureRepository featureRepository) {
+        this.productService = productService;
+        this.featureRepository = featureRepository;
+    }
+
 
     @GetMapping("/all")
-    public ResponseEntity<List<Product>> getAllProduct() {
-        List<Product> product = productService.getAllProduct();
+    public ResponseEntity<MessageDto> getAllProduct() {
+        Set <Product> product = productService.getAllProduct();
         if (product.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageDto(HttpStatus.NOT_FOUND, "No products found."));
         } else {
-            return ResponseEntity.ok(product);
+
+            ResponseEntity<MessageDto> ok = ResponseEntity.ok(new MessageDto(HttpStatus.OK, "Products retrieved successfully.", product));
+            return ok;
         }
     }
 
+
     @GetMapping("/all/")
-    public ResponseEntity<List<Product>> getAllProductFilter(@RequestParam(required = false) String city, @RequestParam(required = false) List<Long> categoriesId, @RequestParam(required = false) List<Long> featuresId) {
-        if (city == null) {
-            city = "";
-        }
-        if (categoriesId == null) {
-            categoriesId = new ArrayList<>();
-        }
-        if (featuresId == null) {
-            featuresId = new ArrayList<>();
-        }
-        List<Product> products = productService.getAllProductFilter(city, categoriesId, featuresId);
+    public ResponseEntity<MessageDto> getAllProductFilter(
+            @RequestParam(defaultValue = "") String city,
+            @RequestParam(defaultValue = "") Set<Long> categoriesId,
+            @RequestParam(defaultValue = "") Set<Long> featuresId) {
+
+        Set<Product> products = (Set<Product>) productService.getAllProductFilter(city, categoriesId, featuresId);
+
         if (products.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageDto(HttpStatus.NOT_FOUND, "No products found."));
         } else {
-            return ResponseEntity.ok(products);
+            return ResponseEntity.ok(new MessageDto(HttpStatus.OK, "Products retrieved successfully.", products));
         }
     }
+
 
     @GetMapping("/productByCategoryId/")
     public ResponseEntity<List<Product>> getProductsByCategoryId(@RequestParam List<Long> categoriesId) {
@@ -80,8 +86,8 @@ public class ProductController {
 
 
     @GetMapping("/productByFeaturesId/")
-    public ResponseEntity<List<Product>> getProductsByFeaturesId(@RequestParam List<Long> featuresId) {
-        List<Product> products = productService.getProductByFeaturesId(featuresId);
+    public ResponseEntity<Set<Product>> getProductsByFeaturesId(@RequestParam Set<Long> featuresId) {
+        Set<Product> products = (Set<Product>) productService.getProductByFeaturesId(featuresId);
         if (products.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         } else {
@@ -90,8 +96,8 @@ public class ProductController {
     }
 
     @GetMapping("/productByFeaturesName/")
-    public ResponseEntity<List<Product>> getProductsByFeaturesName(@RequestParam List<String> featuresName) {
-        List<Product> products = productService.getProductByFeaturesName(featuresName);
+    public ResponseEntity<Set<Product>> getProductsByFeaturesName(@RequestParam Set<String> featuresName) {
+        Set<Product> products = productService.getProductByFeaturesName(featuresName);
         if (products.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         } else {
@@ -108,8 +114,8 @@ public class ProductController {
     }
 
     @GetMapping("/available")
-    public ResponseEntity<List<Product>> getAvailableProduct() throws AttributeException {
-        List<Product> product = productService.getAvailableProduct();
+    public ResponseEntity<Set<Product>> getAvailableProduct() throws AttributeException {
+        Set<Product> product = productService.getAvailableProduct();
         if (product.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         } else {
@@ -119,14 +125,14 @@ public class ProductController {
 
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById (@PathVariable Long id) {
-        Product product = productService.getProductById(id);
-        if (product == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        } else {
+    public ResponseEntity<MessageDto> getProductById(@PathVariable Long id) {
+        Optional<Product> productOptional = Optional.ofNullable(productService.getProductById(id));
 
-            return ResponseEntity.ok(product);
+        if (productOptional.isPresent()) {
+            Product product = productOptional.get();
+            return ResponseEntity.ok(new MessageDto(HttpStatus.OK, "Product retrieved successfully.", product));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageDto(HttpStatus.NOT_FOUND, "Product not found"));
         }
     }
 
@@ -137,7 +143,7 @@ public class ProductController {
         return ResponseEntity.ok(new MessageDto(HttpStatus.OK, message));
     }
 
-    @PutMapping("/update/{id}")
+    @PatchMapping("/update/{id}")
     public ResponseEntity<MessageDto> updateProduct(@PathVariable Long id, @RequestBody ProductDto dto) throws AttributeException {
         productService.updateProduct(id, dto);
 
@@ -145,15 +151,18 @@ public class ProductController {
         return ResponseEntity.ok(new MessageDto(HttpStatus.OK, message));
     }
 
-    @DeleteMapping("/delete/{id}")
+    @PatchMapping("/delete/{id}")
     public ResponseEntity<MessageDto> deleteProduct(@PathVariable Long id) {
 
+        if(!productService.existsById(id)){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageDto(HttpStatus.NOT_FOUND, "Product not found"));
+        }
         productService.deleteProduct(id);
         return ResponseEntity.ok(new MessageDto(HttpStatus.OK, "Product deleted successfully"));
     }
 
     @GetMapping("/urlImage/{id}")
-    public List<Map<String, Object>> getProducttModifiedImages(@PathVariable Long id) {
+    public Set<Map<String, Object>> getProducttModifiedImages(@PathVariable Long id) {
         return productService.findProductWithModifiedImages(id);
     }
